Refactor and Reorganize the code for following Tests getEmptyInvoiceTest() getInvoiceTest()
