+++ src/main/java/com/hundred/percent/capstone/Invoicify/company/controller/CompanyController.java
@@ -1,10 +1,7 @@
 package com.hundred.percent.capstone.Invoicify.company.controller;
 
 import com.hundred.percent.capstone.Invoicify.company.dto.CompanyDTO;
-import com.hundred.percent.capstone.Invoicify.company.dto.InvoiceDTO;
-import com.hundred.percent.capstone.Invoicify.company.entity.CompanyEntity;
 import com.hundred.percent.capstone.Invoicify.company.service.CompanyService;
-import com.hundred.percent.capstone.Invoicify.company.service.InvoiceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
